outsource handleSubmittableEvents and apiPromise
